solve request conflicts problem between rails and react with an api namespace
