+++ package.json
@@ -4,6 +4,7 @@
   "dependencies": {
     "@babel/preset-react": "^7.8.3",
     "@rails/webpacker": "4.2.2",
+    "axios": "^0.19.2",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "eslint-plugin-react": "^7.18.3",
     "preact": "^10.3.2",
